+++ server/src/main/java/org/allaymc/server/entity/component/projectile/EntityProjectileBaseComponentImpl.java
@@ -52,11 +52,6 @@ public class EntityProjectileBaseComponentImpl extends EntityBaseComponentImpl i
     }
 
     @Override
-    public double getStepHeight() {
-        return 0;
-    }
-
-    @Override
     public boolean computeEntityCollisionMotion() {
         return false;
     }